moved getPalette to ColorProvider
