+++ src/components/color/ColorProvider.js
@@ -32,6 +32,12 @@ export const ColorProvider = (props) => {
     return authFetch(path, options).then((res) => res.json());
   };
 
+  const getPalette = (name) => {
+    const path =
+      name === "default" ? "/default/palette" : `/palettes?name=${name}`;
+    return authFetch(path, { noAuth: true }).then((res) => res.json());
+  };
+
   return (
     <ColorContext.Provider
       value={{
@@ -39,6 +45,7 @@ export const ColorProvider = (props) => {
         getDefaultColors,
         color,
         setColor,
+        getPalette,
       }}
     >
       {props.children}
